Rename packages and update license statements
